+++ packages/thumbprint-react/components/Modal/components/sticky-footer.tsx
@@ -1,6 +1,10 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import styles from './sticky-footer.module.scss';
+
+interface PropTypes {
+    children: React.ReactNode;
+    setSticky: (isSticky: boolean) => void;
+}
 
 /**
  * This component exists because `this.props.setSticky` updates state and React doesn't allow state
@@ -8,26 +12,15 @@ import styles from './sticky-footer.module.scss';
  * (with the Context API), so it's not possible to access the function within its lifecycle hooks.
  * https://blog.kentcdodds.com/answers-to-common-questions-about-render-props-a9f84bb12d5d#6a05
  */
-class StickyFooter extends React.Component {
-    componentDidMount() {
+export default class StickyFooter extends React.Component<PropTypes> {
+    componentDidMount(): void {
         const { setSticky } = this.props;
         setSticky(true);
     }
 
-    render() {
+    render(): JSX.Element {
         const { children } = this.props;
 
         return <div className={styles.root}>{children}</div>;
     }
 }
-
-StickyFooter.propTypes = {
-    children: PropTypes.node.isRequired,
-    setSticky: PropTypes.func,
-};
-
-StickyFooter.defaultProps = {
-    setSticky: undefined,
-};
-
-export default StickyFooter;